refactor(lib/xchain): rename xtypes to xchain
