+++ lib/xchain/doc.go
@@ -1,0 +1,2 @@
+// Package xchain defines the types and interfaces used by the omni cross-chain protocol.
+package xchain
